-added a scenarios directory and moved Scenarios.vue into it -added a route.ts model interface to define the shape of TopNavbar.vue routes -added typing to route list in TopNavbar.vue
