+++ BridgeCareApp/VuejsApp/src/components/scenarios/Scenarios.vue
@@ -81,8 +81,8 @@
     import axios from 'axios';
 
     import * as moment from 'moment';
-    import {Scenario} from '@/shared/models/iAM/scenario';
-    import {hasValue} from '@/shared/utils/has-value';
+    import {Scenario} from '../../shared/models/iAM/scenario';
+    import {hasValue} from '../../shared/utils/has-value';
 
     axios.defaults.baseURL = process.env.VUE_APP_URL;
