+++ BridgeCareApp/VuejsApp/src/shared/models/iAM/route.ts
@@ -1,0 +1,6 @@
+export interface Route {
+    navigation: string;
+    icon: string;
+    name: string;
+    linkGroup: string;
+}
